Changed "state-bred" to Restriction -> Only showing Province/State not country
Weeks Page > when building TEL, races are default sorted by number of confirmed horses

When displaying the TEL (Published or Unpublished) for any user, the Priority Races are always shown on top sorted by field size, followed by all other races sorted by field size

Race Menu: If a 2-yo, sex options are Open and C/G. If a 2-yo Filly, sex options are F and Open. If 3-yo, sex options are Open and C/G. If a 3-yo-filly and for 3-yo races only, sex options are F and Open. If a 3-yo-filly and the races are for 3yo+, then the sex options are F/M and Open. If a 4-yo, sex options are Open and C/G. If a 4-yo filly and for 4-yo races only, sex options are F and Open. If a 4-yo filly and the races are for 4yo+, then the sex options are F/M and Open. For each of these cases, the default should be set to whatever the first choice I listed here.

If the horse is a female horse, than sex will default to filly and/or mare (other option open)
If the horse is a male horse, than sex will default to open
If the horse is a maiden, than conditions will default to maiden
Once a trainer has begun to use the filtering options, those options will save until either the trainer has gone in and changed them again or when the race office has updated the horse’s newest conditions.
If the race office has updated the horse’s newest conditions, the race menu will automatically be reset back to a blank in each category.

Race > Has other conditions flag to indicate race has a non screenable conditions, displayed on race menu

When creating a race, that race no longer needs to be saved before adding conditions

EL-> When TEL becomes EL a reset instance of race becomes available in TEL database
	-> EL cannot be converted back to TEL
	-> confirmation pop up for converting TEL
	-> EL not shown on home page
	-> Horses on an EL are frozen, status set to 'Running'
	-> Button on race finished view to reset all horses to race ready (until race day is added)
	-> Scratching horse also sets status back to Race Ready