Name field disabled for the trainer/owner on the "My Account Page"
Bug Fix: Unrecognized characters for race purses now handled.

Pop up to confirm setting a race back to draft on races page

When a horse is already confirmed for a race and tried to confirm to another, a pop up will be displayed saying "This horse is already confirmed to *RACE_NAME*. Would you like to unconfirm from *RACE_NAME* and confirm to this race?"

Priority Column now showing dates for specified tables.

Trainer/Owner can now confirmed eligible horses in a published TEL on the homepage. (wanting to confirm whether i should add horse who are confirmed to a different race and show same confirm popup or leave confirmed horses out of the list)