Once a trainer has begun to use the filtering options, those options will save until either the trainer has gone in and changed them again or when the race office has updated the horse’s newest conditions. If the race office has updated the horse’s newest conditions, the race menu will automatically be reset back to a blank in each category. Otherwise they should remain set to last chosen combination, even if the user leaves the page, logs out, etc.
the race menu settings should remain the same after a horse confirms to a race.

"Conditions" filter needs to have an open and an all. Open refers to races with no conditions. All refers to races with and without conditions.

Added confirm modal to homepage for adding horse to TEL.

Need a new date field in the “Create Race” form called “Nominations Close”. This will be displayed on the Stakes Schedule table only (see excel sheet called  "Tables")

Fix bug with race date not being duplicated